requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.13.4
billiard==4.2.1
boto3==1.35.22
botocore==1.35.22
CacheControl==0.14.1
cachetools==5.5.0
celery==5.4.0
certifi==2024.2.2
cffi==1.17.1
channels==4.2.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==43.0.1
decouple==0.0.7
deep-translator==1.11.4
dj-rest-auth==6.0.0
Django==5.0.4
django-allauth==64.2.0
django-cors-headers==4.3.1
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-filter==24.3
django-storages==1.14.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
firebase-admin==6.6.0
frozenlist==1.4.1
google-api-core==2.23.0
google-api-python-client==2.153.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
googletrans==4.0.0rc1
grpcio==1.68.0
grpcio-status==1.68.0
gunicorn==23.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httplib2==0.22.0
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
jmespath==1.0.1
joblib==1.4.2
kombu==5.4.2
langdetect==1.0.9
msgpack==1.1.0
multidict==6.0.5
nltk==3.9.1
numpy==2.2.4
packaging==24.1
pandas==2.2.3
phonenumbers==8.13.49
pillow==10.3.0
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.28.3
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
PyJWT==2.8.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
qrcode==8.0
regex==2024.11.6
requests==2.31.0
rfc3986==1.5.0
rsa==4.9
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.7
sqlparse==0.5.0
textblob==0.19.0
tqdm==4.67.1
twilio==9.2.3
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
XlsxWriter==3.2.2
yarl==1.9.4
